lambda/requirements.txt
boto3==1.34.16
botocore==1.34.16
pycryptodome==3.20.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.7
